+++ behave/features/environment.py
@@ -2,14 +2,13 @@ from fixtures.helper import helper
 from time import sleep
 import json
 
-def before_all(context):
-
-    context.entity_to_delete = {}
-
 def before_scenario(context, scenario):
 
     if "network" in scenario.tags:
         context.arr_network_to_delete = []
+
+    context.entity_to_delete = {}
+    context.result = {}
         
 def after_scenario(context, scenario):
     
@@ -19,90 +18,66 @@ def after_scenario(context, scenario):
         entity_plural = "virtual_machines"
         entity_singular = "virtual_machine"
 
-        if "cold_migration" in context.feature.tags and "linux" in context.feature.tags and "statichv" in context.feature.tags:
-            name = "linux-vm-without-startup-static"
+        for identifier in context.entity_to_delete["onapp_vm"]:
+            data = context.cp.search(entity_plural, args=identifier)
 
-        elif "cold_migration" in context.feature.tags and "windows" in context.feature.tags and "statichv" in context.feature.tags:
-            name = "windows-vm-without-startup-static"
+            if data:
+                id = data[0][entity_singular]["id"]
 
-        elif "hot_migration" in context.feature.tags and "linux" in context.feature.tags and "statichv" in context.feature.tags:
-            name = "linux-vm-with-startup-static"
+                # check whether vm is in suspended state before delete
+                if data[0][entity_singular]["suspended"]:
+                    # activate it back by sending 'suspend' action again
+                    suspension_response = context.cp.post_action(entity_plural, id, "suspend")
 
-        elif "hot_migration" in context.feature.tags and "windows" in context.feature.tags and "statichv" in context.feature.tags:
-            name = "windows-vm-with-startup-static"
+                    if suspension_response.status_code != 201:
+                        assert False, "error: unable to activate the VM (%s) in onapp cloud" % data[0][entity_singular]["label"]
 
-        elif "cold_migration" in context.feature.tags and "linux" in context.feature.tags and "cloudboothv" in context.feature.tags:
-            name = "linux-vm-without-startup-cloudboot"
+                print("VM (%s) found in Onapp cloud, proceed to delete..." % data[0][entity_singular]["label"])
 
-        elif "cold_migration" in context.feature.tags and "windows" in context.feature.tags and "cloudboothv" in context.feature.tags:
-            name = "windows-vm-without-startup-cloudboot"
+                response = {}
+                response[entity_plural] = context.cp.delete(entity_plural, id)
 
-        elif "hot_migration" in context.feature.tags and "linux" in context.feature.tags and "cloudboothv" in context.feature.tags:
-            name = "linux-vm-with-startup-cloudboot"
+                sleep(60)
 
-        elif "hot_migration" in context.feature.tags and "windows" in context.feature.tags and "cloudboothv" in context.feature.tags:
-            name = "windows-vm-with-startup-cloudboot"
+                if response[entity_plural].status_code == 204:
+                    print("VM has been deleted successfully")
+                else:
+                    print("error: failed to delete, {error}".format(error=response[entity_plural].text))
 
-        fixture = helper.get_fixture(entity_singular)[name][entity_singular]
-        label = fixture["label"]
-        data = context.cp.search(entity_plural, args=label)
-
-        if data:
-            id = data[0][entity_singular]["id"]
-
-            # check whether vm is in suspended state before delete
-            if data[0][entity_singular]["suspended"]:
-                # activate it back by sending 'suspend' action again
-                suspension_response = context.cp.post_action(entity_plural, id, "suspend")
-
-                if suspension_response.status_code != 201:
-                    assert False, "error: unable to activate the VM (%s) in onapp cloud" % label
-
-            print("VM found in Onapp cloud, proceed to delete...")
-
-            response = {}
-            response[entity_plural] = context.cp.delete(entity_plural, id)
-
-            sleep(60)
-
-            if response[entity_plural].status_code == 204:
-                print("VM has been deleted successfully")
             else:
-                print("error: failed to delete, {error}".format(error=response[entity_plural].text))
-
-        else:
-            print("VM is not found, proceed to delete the VM in VHI portal")
+                print("VM is not found, proceed to delete the VM in VHI portal")
 
         # VHI portal
-        hostname = fixture["hostname"]
-        config = helper.get_config()
-        output = helper.open_vhi_ssh_connection(config["vhi"], "service compute server list -f json")
-        vm_list = json.loads(output.stdout)
-
-        match = False
-        for vm in vm_list:
-
-            if hostname in vm["name"]:
-                match = True
-
-                # find the volume from server
-                volume_output = helper.open_vhi_ssh_connection(config["vhi"], "service compute server volume list --server %s -f json" % vm["name"])
-                arr_volume = json.loads(volume_output.stdout)
-
-                arr_device = []
-                for device in arr_volume:
-                    arr_device.append(device["id"])
-                
-                # to delete vm in vhi portal
-                print("VM found in VHI portal, proceed to delete...")
-                _ = helper.open_vhi_ssh_connection(config["vhi"], "service compute server delete {vm_name}".format(vm_name=vm["name"]))
-
-                print("VM has been deleted successfully")
-                break
+        for hostname in context.entity_to_delete["vhi_vm"]:
         
-        # we proceed with the rest of the scenario even if the vm is not found
-        if not match:
-            print("VM is not found in VHI portal, proceed to next scenario")
+            config = helper.get_config()
+            output = helper.open_vhi_ssh_connection(config["vhi"], "service compute server list -f json")
+            vm_list = json.loads(output.stdout)
+
+            match = False
+            for vm in vm_list:
+
+                if hostname in vm["name"]:
+                    match = True
+
+                    # find the volume from server
+                    volume_output = helper.open_vhi_ssh_connection(config["vhi"], "service compute server volume list --server %s -f json" % vm["name"])
+                    arr_volume = json.loads(volume_output.stdout)
+
+                    arr_device = []
+                    for device in arr_volume:
+                        arr_device.append(device["id"])
+                    
+                    # to delete vm in vhi portal
+                    print("VM ({vm_name}) found in VHI portal, proceed to delete...".format(vm_name=vm["name"]))
+                    _ = helper.open_vhi_ssh_connection(config["vhi"], "service compute server delete {vm_name}".format(vm_name=vm["name"]))
+
+                    print("VM has been deleted successfully")
+                    break
+            
+                # we proceed with the rest of the scenario even if the vm is not found
+                if not match:
+                    print("VM is not found in VHI portal, proceed to next scenario")
         
         if "arr_device" in locals():
             for id in arr_device:
@@ -143,12 +118,10 @@ def after_scenario(context, scenario):
                         match = True
                         break
 
-            if not match:
-                assert False, "error: flavor is not found in placement"
-            
-            _ = helper.open_vhi_ssh_connection(config["vhi"], "service compute placement delete-assign --flavor {flavor} {name}"\
-                                                    .format(flavor=flavor_name, name=name))
-            print("flavor named %s has been unassigned from placement %s" % (flavor_name, name))
+            if match:
+                _ = helper.open_vhi_ssh_connection(config["vhi"], "service compute placement delete-assign --flavor {flavor} {name}"\
+                                                        .format(flavor=flavor_name, name=name))
+                print("flavor named %s has been unassigned from placement %s" % (flavor_name, name))
 
             _ = helper.open_vhi_ssh_connection(config["vhi"], "service compute placement delete {name}".format(name=name))
             print("placement named %s has been removed" % name)