+++ onapp2vhi/inc/onapp_helpers.py
@@ -395,6 +395,7 @@ def get_vm_source_properties(cfg: OnApp2VHIConfig, vm_idn: str) -> Dict:
     :param vm_idn:
     :return:
     """
+    version = onapp_version(cfg)
     onapp_requests = OnAppRequests(cfg)
     vm_properties = onapp_requests.get(f'virtual_machines/{vm_idn}')['virtual_machine']
     network_info = get_vm_network_info(cfg, vm_identifier=vm_idn)
@@ -406,9 +407,14 @@ def get_vm_source_properties(cfg: OnApp2VHIConfig, vm_idn: str) -> Dict:
     _hv_props = onapp_requests.get(f'settings/hypervisors/{_vm_hv_id}')
     _vm_hv_ip = _hv_props['hypervisor']['ip_address']
     _vm_nics = onapp_requests.get(f'virtual_machines/{vm_idn}/ip_addresses')
-    _vm_ip_addr = [nic['ip_address_join']['ip_address']['address'] for nic in _vm_nics
-                   if (nic['ip_address_join']['ip_address']
-                       and nic['ip_address_join']['ip_address']['primary'])][0]
+    if version >= 6.4:
+        _vm_ip_addr = [nic['ip_address_join']['ip_address']['address'] for nic in _vm_nics
+                       if (nic['ip_address_join']['ip_address']
+                           and nic['ip_address_join']['ip_address']['primary'])][0]
+    else:
+        _vm_ip_addr = [nic['ip_address_join']['ip_address']['address'] for nic in _vm_nics
+                       if (nic['ip_address_join']['ip_address']
+                           and nic['ip_address_join']['ip_address']['address'])][0]
     logs.info(f"-- Hypervisor ID: {_vm_hv_id} | Hypervisor IP ADDRESS: {_vm_hv_ip} | VM IP ADDRESS {_vm_ip_addr}")
     return {'hv_ip': _vm_hv_ip, 'vm_os': _vm_os, 'vm_ip_addr': _vm_ip_addr, 'network_info': network_info,
             'hot_migrate': _hot_migrate, 'hostname': _vm_hostname, 'domain': _vm_domain}
@@ -746,6 +752,7 @@ class VmHandler:
                 return vm_install_bootloader_offline, vm_cold_migrate
 
 
+# Todo: Move this class to separate module
 class GenerateXmlConfig:
     RECOVERY_TEMPLATE = 'ls /onapp/tools/recovery/recovery-centos-7.*.{file} | tail -1'
 
@@ -771,7 +778,7 @@ class GenerateXmlConfig:
         :return:
         """
         cmd_1 = (f"virsh dumpxml {self._vm_idn} 2>/dev/null > /tmp/{self._vm_idn}.xml;"
-                 f" cat /tmp/{self._vm_idn}.xml' 2>/dev/null")
+                 f" cat /tmp/{self._vm_idn}.xml 2>/dev/null")
         exit_status, vm_xml_cfg = self.hv_ssh.execute(command=cmd_1)
         vm_xml = KVMxml.fromstring(vm_xml_cfg)
         for device in vm_xml.findall("devices"):